requirements.txt
# Core
py-clob-client==0.17.5
web3==6.20.1
websockets==12.0
aiohttp==3.9.5
python-dotenv==1.0.1

# Utility
pydantic==2.7.1
rich==13.7.1
python-dateutil==2.9.0

# Telegram (optional but default-on in code)
aiogram==3.7.0

# Math
numpy==1.26.4
